refactor: 거리계산하는 MapUtils 클래스 역할을 Position 객체로 위임
- 사용하지 않는 MapUtils 제거
