+++ BE/src/main/java/team13/kuje/airbnb/domain/Place.java
@@ -9,7 +9,6 @@ import javax.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import team13.kuje.airbnb.util.MapUtils;
 
 @Entity
 @Getter
@@ -32,8 +31,7 @@ public class Place {
 
 
 	public int calculateTime(Position inputPosition) {
-		double distance = MapUtils.calculateDistance(this.position.getLat(), this.position.getLng(),
-			inputPosition.getLat(), inputPosition.getLng());
+		double distance = position.calculateDistance(inputPosition);
 
 		return (int) ((distance / AVERAGE_SPEED) * HOUR_TO_MIN);
 	}
